% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_recipe.R
\name{adae_sjf}
\alias{adae_sjf}
\alias{adae,}
\alias{ADAE}
\alias{aes_func}
\title{ADAE recipe helper functions and constants}
\usage{
adae_sjf(n, .db, .df)

aes_func(n, .df, lookup = adae_lookup)
}
\arguments{
\item{n}{numeric(1). Number of observations to generate. Ignored if \code{.df} is specified}

\item{.db}{list. Named list of data.frames for use in function when scaffolding}

\item{.df}{data.frame. Partially generated data table to build upon}

\item{lookup}{data.frame. Data frame with possible adverse-event information to be sampled from.}
}
\description{
ADAE recipe helper functions and constants
}
