% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeg_recipe.R
\docType{data}
\name{adae_scaff}
\alias{adae_scaff}
\alias{adeg_scaff_recipe}
\alias{adeg_tbl_recipe}
\title{Recipes for creating ADEG CDISC Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 4 columns.
}
\usage{
adae_scaff(n, .db, .df)

adeg_scaff_recipe

adeg_tbl_recipe
}
\description{
Recipes for creating ADEG CDISC Data
}
\keyword{datasets}
