% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_recipe.R
\docType{data}
\name{s_countries}
\alias{s_countries}
\alias{country_site_prob}
\alias{s_armcds}
\alias{s_armprobs}
\alias{s_sex}
\alias{s_sexprobs}
\alias{s_race}
\alias{s_raceprobs}
\alias{rchisq_age}
\alias{s_strat1}
\alias{s_strat2}
\alias{s_bmrkr2}
\alias{trtdtm_varnames}
\alias{sample_trtdtmvars}
\alias{eos_varnames}
\alias{s_discon}
\alias{make_eosvars}
\alias{sample_siteid}
\alias{arm_varnames}
\alias{sample_armcd}
\alias{usubj_deps}
\alias{usubj_vars}
\title{Helper functions and constants for ADSL}
\format{
An object of class \code{character} of length 10.

An object of class \code{numeric} of length 10.

An object of class \code{character} of length 3.

An object of class \code{numeric} of length 3.

An object of class \code{character} of length 4.

An object of class \code{numeric} of length 4.

An object of class \code{character} of length 8.

An object of class \code{numeric} of length 8.

An object of class \code{character} of length 3.

An object of class \code{character} of length 2.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 4.

An object of class \code{character} of length 6.

An object of class \code{character} of length 4.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.
}
\usage{
s_countries

country_site_prob

s_armcds

s_armprobs

s_sex

s_sexprobs

s_race

s_raceprobs

rchisq_age(n)

s_strat1

s_strat2

s_bmrkr2

trtdtm_varnames

sample_trtdtmvars(
  n,
  study_duration = 2,
  sys_dtm = as.numeric(strptime("20/2/2019 11:16:16.683", "\%d/\%m/\%Y \%H:\%M:\%OS")),
  discons = floor(n * 0.3)
)

eos_varnames

s_discon

make_eosvars(.df, n = NROW(.df))

sample_siteid(.df, n, x, prob)

arm_varnames

sample_armcd(
  n,
  narms = 3,
  armnms = c(`ARM A` = "A: Drug X", `ARM B` = "B: Placebo", `ARM C` = "C: Combination")
)

usubj_deps

usubj_vars
}
\description{
Helper functions and constants for ADSL
}
\keyword{datasets}
