% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_recipe.R
\docType{data}
\name{adsl_recipe}
\alias{adsl_recipe}
\title{Table and Joint Recipes for CDISC datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 4 columns.
}
\usage{
adsl_recipe
}
\description{
Table and Joint Recipes for CDISC datasets
}
\keyword{datasets}
