% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_recipe.R, R/adaette_recipe.R,
%   R/adcm_recipe.R, R/adeg_recipe.R, R/adsl_recipe.R, R/adtte_recipe.R
\docType{data}
\name{adae_scaff_recipe}
\alias{adae_scaff_recipe}
\alias{adae_tbl_recipe}
\alias{adaette_scaff_recipe}
\alias{adaette_tbl_recipe}
\alias{adcm_scaff_recipe}
\alias{adcm_tbl_recipe}
\alias{adae_scaff}
\alias{adeg_scaff_recipe}
\alias{adeg_tbl_recipe}
\alias{adsl_tbl_recipe}
\alias{tte_scaff_recipe}
\alias{tte_tbl_recipe}
\title{Recipes for creating ADEG CDISC Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.
}
\usage{
adae_scaff_recipe

adae_tbl_recipe

adaette_scaff_recipe

adaette_tbl_recipe

adcm_scaff_recipe

adcm_tbl_recipe

adae_scaff(n, .db, .df)

adeg_scaff_recipe

adeg_tbl_recipe

adsl_tbl_recipe

tte_scaff_recipe

tte_tbl_recipe
}
\arguments{
\item{n}{numeric(1). Number of observations to generate. Ignored if \code{.df} is specified}

\item{.db}{list. Named list of data.frames for use in function when scaffolding}

\item{.df}{data.frame. Partially generated data table to build upon}
}
\description{
Subject Level Dataset
}
\details{
This particular recipe implements a simple generic ADSL dataset.
}
\examples{


adsl <- gen_table_data(N = 10, adsl_tbl_recipe)
adae <- gen_reljoin_table(adae_scaff_recipe, adae_tbl_recipe, db = list(ADSL = adsl))
adtte <- gen_reljoin_table(tte_scaff_recipe, tte_tbl_recipe, db = list(ADSL = adsl))

adaette <- gen_reljoin_table(adaette_scaff_recipe, adaette_tbl_recipe,
                             db = list(ADSL = adsl, ADAE = adae))

#TODO adcm <- gen_reljoin_table(acdm_scaff_recipe, adcm_tbl_recipe, db = list(ADSL = adsl))

}
\keyword{datasets}
