% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeg_recipe.R
\name{gen_adeg_seq}
\alias{gen_adeg_seq}
\title{Generate Sequence Per Number of USUBJID Observation}
\usage{
gen_adeg_seq(n, .df, ...)
}
\arguments{
\item{n}{ignored.}

\item{.df}{data frame with required variable `USUBJID`}
}
\description{
Generate Sequence Per Number of USUBJID Observation
}
\examples{
x <- data.frame(USUBJID = rep(1:10, each = 2))

gen_adeg_seq(NULL, x)
gen_adeg_seq(NULL, data.frame(USUBJID = c('id1', 'id1', 'id2', 'id3', 'id3', 'id3')))

}
