% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg_conventions.R
\name{gen_args}
\alias{gen_args}
\title{General Argument Convention}
\usage{
gen_args(n, .df, .db, lookup)
}
\arguments{
\item{n}{numeric(1). Number of observations to generate. Ignored if \code{.df} is specified}

\item{.df}{data.frame. Partially generated data table to build upon}

\item{.db}{list. Named list of data.frames for use in function when scaffolding}

\item{lookup}{data.frame. Table of lookup values specific to the function at hand.}
}
\value{
Not important. function is an argument template
}
\description{
General Argument Convention
}
