% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adeg_recipe.R
\name{visit_schedule}
\alias{visit_schedule}
\title{Create visit schedule}
\usage{
visit_schedule(
  visit_format = "WEEK",
  n_assessments = 10L,
  n_days = 5L,
  add_min_max = TRUE
)
}
\arguments{
\item{visit_format}{as character string. Valid values: WEEK, CYCLE.}

\item{n_assessments}{number of assessments. Valid values: integer.}

\item{n_days}{number of days for each cycle: Valid values: integer.}

\item{add_min_max}{logical(1). Should the min/max post-baseline values be added.}
}
\value{
a factor of length n_assessments
}
\description{
Create a visit schedule as factor.

X number of visits or X number of cycles and Y number of days.
}
\examples{
visit_schedule(visit_format = "WEeK", n_assessments = 10L)
visit_schedule(visit_format = "CyCLE", n_assessments = 5L, n_days = 2L)
}
