% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaette_recipe.R
\docType{data}
\name{adaette_fdeps}
\alias{adaette_fdeps}
\alias{adaette_sjvars}
\alias{adaette_paramcds}
\alias{fullcensor}
\alias{adaette_one_pat}
\alias{adaette_sjf}
\alias{evntdescr_sel}
\alias{cnsdtdscr_sel}
\alias{adaette_cnsrdtdsc}
\title{ADAETTE recipe helper constants and functions}
\format{
An object of class \code{character} of length 5.

An object of class \code{character} of length 5.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 3 columns.

An object of class \code{character} of length 5.

An object of class \code{character} of length 4.
}
\usage{
adaette_fdeps

adaette_sjvars

adaette_paramcds

fullcensor(usubjid, aereptte, adtm, ady)

adaette_one_pat(pat_info, s_dur_secs = pat_info$study_duration_secs)

adaette_sjf(.db, lookup = adaette_paramcds)

evntdescr_sel

cnsdtdscr_sel

adaette_cnsrdtdsc(n, .df, evnt_sel = evntdescr_sel, cns_sel = cnsdtdscr_sel)
}
\arguments{
\item{usubjid}{character(1). USUBJID}

\item{aereptte}{numeric(1). Time to end of AE reporting in years.}

\item{adtm}{POSIXct(1). Analysis datetime}

\item{ady}{numeric(1). Relative day number of analysis time}

\item{pat_info}{data.frame. ADAE data for a single subjid}

\item{s_dur_secs}{numeric(1). Total study duration in seconds}

\item{.db}{list. Named list of data.frames for use in function when scaffolding}

\item{lookup}{data.frame. Table of lookup values specific to the function at hand.}

\item{n}{numeric(1). Number of observations to generate. Ignored if \code{.df} is specified}

\item{.df}{data.frame. Partially generated data table to build upon}

\item{evnt_sel}{character. Event descriptions to sample from.}

\item{cns_sel}{character. Reasons for censoring to sample from.}
}
\description{
ADAETTE recipe helper constants and functions
}
\keyword{datasets}
