% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adsl_recipe.R
\docType{data}
\name{adsl_recipe}
\alias{adsl_recipe}
\title{Recipe for the ADSL dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 4 columns.
}
\usage{
adsl_recipe
}
\description{
Subject Level Dataset
}
\details{
This particular recipe implements a simple generic ADSL dataset.
}
\examples{

adsl_recipe

gen_table_data(N = 10, recipe = adsl_recipe)

}
\keyword{datasets}
