% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adtte_recipe.R
\docType{data}
\name{tte_rel_join_recipe}
\alias{tte_rel_join_recipe}
\alias{tte_table_recipe}
\title{Recipes for creating ADTTE CDISC Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 4 columns.
}
\usage{
tte_rel_join_recipe

tte_table_recipe
}
\description{
Recipes for creating ADTTE CDISC Data
}
\examples{

ADSL <- gen_table_data(N = 10, recipe = adsl_recipe)
gen_reljoin_table(tte_rel_join_recipe, tte_table_recipe, db = list(ADSL = ADSL))

}
\keyword{datasets}
