% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adae_recipe.R, R/adaette_recipe.R,
%   R/adsl_recipe.R
\docType{data}
\name{adae_scaff_recipe}
\alias{adae_scaff_recipe}
\alias{adae_tbl_recipe}
\alias{adaette_scaff_recipe}
\alias{adaette_tbl_recipe}
\alias{adsl_tbl_recipe}
\title{Recipe for the ADSL dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 5 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 4 columns.
}
\usage{
adae_scaff_recipe

adae_tbl_recipe

adaette_scaff_recipe

adaette_tbl_recipe

adsl_tbl_recipe
}
\description{
Subject Level Dataset
}
\details{
This particular recipe implements a simple generic ADSL dataset.
}
\examples{

adaette <- gen_reljoin_table(adaette_sjrec, adaette_recipe,
                             db = list(ADSL = adsl, ADAE = adae))

adsl_recipe

adsl <- gen_table_data(N = 10, adsl_recipe)

}
\keyword{datasets}
